refactor(xtrellis): extract args and coordinator to separate files
